Refactor tests structure
* Move functional tests under the tests package.
* Moved demos outside of the e2e package and into the project folder.
* Moved driver outside of the e2e package and into the project folder.
* Moved pages outside of the e2e package and into the project folder and renamed it to page.
* Update demo.sh script.
