--- e2e/driver/__init__.py
@@ -1,2 +1,0 @@
-from e2e.driver.selenium_driver import SeleniumDriver
-from e2e.driver.seleniumbase_driver import SeleniumBaseDriver
