+++ page/__init__.py
@@ -1,0 +1,6 @@
+from page.admin import AdminLoginPage
+from page.drafts import DraftsPage
+from page.home import HomePage
+from page.new_comment import NewCommentPage
+from page.new_post import NewPostPage
+from page.post import PostPage
