+++ tests/functional/test_post_create.py
@@ -3,14 +3,14 @@ from django.test import LiveServerTestCase
 from selenium.webdriver.firefox import options
 from selenium.webdriver.firefox.webdriver import WebDriver
 
-from e2e.driver import SeleniumDriver
-from e2e.pages.admin import AdminLoginPage
-from e2e.pages.home import HomePage
-from e2e.pages.new_post import NewPostPage
-from e2e.pages.post import PostPage
+from driver import SeleniumDriver
+from page import AdminLoginPage
+from page import DraftsPage
+from page import HomePage
+from page import NewPostPage
 
 
-class PostDeleteTestCase(LiveServerTestCase):
+class PostCreateTestCase(LiveServerTestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
@@ -27,10 +27,10 @@ class PostDeleteTestCase(LiveServerTestCase):
         self.page_driver = SeleniumDriver(self.webdriver)
         get_user_model().objects.create_superuser('admin', 'admin@example.com', 'password')
 
-    def test_a_post_can_be_deleted(self):
+    def test_a_post_can_be_created(self):
         home_page = HomePage(self.page_driver, self.live_server_url)
         new_post_page = NewPostPage(self.page_driver)
-        post_page = PostPage(self.page_driver)
+        drafts_page = DraftsPage(self.page_driver, self.live_server_url)
 
         self.login()
 
@@ -39,15 +39,8 @@ class PostDeleteTestCase(LiveServerTestCase):
 
         post_title = 'Post Title'
         new_post_page.add_post(post_title, 'Post content')
-        new_post_page.publish_post()
 
-        home_page.load()
-        home_page.view_post(post_title)
-
-        post_page.delete_post()
-
-        home_page.load()
-        self.assertFalse(home_page.has_post_with_title(post_title))
+        self.assertTrue(drafts_page.has_post_with_title(post_title))
 
         home_page.logout()
