+++ tests/functional/test_post_publish.py
@@ -3,14 +3,13 @@ from django.test import LiveServerTestCase
 from selenium.webdriver.firefox import options
 from selenium.webdriver.firefox.webdriver import WebDriver
 
-from e2e.driver import SeleniumDriver
-from e2e.pages.admin import AdminLoginPage
-from e2e.pages.drafts import DraftsPage
-from e2e.pages.home import HomePage
-from e2e.pages.new_post import NewPostPage
+from driver import SeleniumDriver
+from page import AdminLoginPage
+from page import HomePage
+from page import NewPostPage
 
 
-class PostCreateTestCase(LiveServerTestCase):
+class PostPublishTestCase(LiveServerTestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
@@ -20,17 +19,16 @@ class PostCreateTestCase(LiveServerTestCase):
 
     @classmethod
     def tearDownClass(cls):
-        super().tearDownClass()
         cls.webdriver.quit()
+        super().tearDownClass()
 
     def setUp(self):
         self.page_driver = SeleniumDriver(self.webdriver)
         get_user_model().objects.create_superuser('admin', 'admin@example.com', 'password')
 
-    def test_a_post_can_be_created(self):
+    def test_a_post_can_be_published(self):
         home_page = HomePage(self.page_driver, self.live_server_url)
         new_post_page = NewPostPage(self.page_driver)
-        drafts_page = DraftsPage(self.page_driver, self.live_server_url)
 
         self.login()
 
@@ -39,8 +37,10 @@ class PostCreateTestCase(LiveServerTestCase):
 
         post_title = 'Post Title'
         new_post_page.add_post(post_title, 'Post content')
+        new_post_page.publish_post()
 
-        self.assertTrue(drafts_page.has_post_with_title(post_title))
+        home_page.load()
+        self.assertTrue(home_page.has_post_with_title(post_title))
 
         home_page.logout()
